heimdal: switch to master branch, use trixie only
The "synthetic-principals" branch we were using would not build cleanly on
'trixie' and required first building everything for 'bookworm'. The master
branch doesn't have this problem.

Nonetheless, this also adds some of the build smarts to be able to build
multiple versions at the same time, including the case of building on a
different distro to the one being installed on.

Signed-off-by: Geoffrey Thorpe <geoff@geoffthorpe.net>