kdcsvc: enforce HTTPS with client-authentication
It makes no sense for the KDC admin API to be optional about HTTPS nor
about client-auth, and having it mandatory allows us to depend on the
presence of credential information. Here, the reference use case restricts
use of the API to a successfully-authenticated client certificate
containing the "orchestrator@hcphacking.xyz" email address.

Signed-off-by: Geoffrey Thorpe <geoff@geoffthorpe.net>