python/HcpToolOrchestrator: add "-a" to avoid querying enrollsvc
If we know the status of TPM creation and enrollment for the hosts we're
interested in, we can use '-a' to bypass orchestrator's default behavior
of querying the enrollsvc to discover status. This is especially useful if
enrollsvc isn't running, which will be the case when we allow enrollsvc to
be attestsvc-based. (We'll need to create the TPM and inject it into the
database before enrollsvc can run.)

Signed-off-by: Geoffrey Thorpe <geoff@geoffthorpe.net>