hcp/attestsvc: add experimentation for attestsvc hook
In addition to the quote-checking and other discipline handled by the
attestsvc front-end, the backend should also have a hook to look at the
attesting host's PCR values and decide if it accepts them. It's not clear
yet how to bend tpm2 to this will, but this code helps with the
exploration.

Signed-off-by: Geoffrey Thorpe <geoff@geoffthorpe.net>