hcp: fix ambiguity about HCP_CONFIG_FILE
When launched, the container's environment has a HCP_CONFIG_FILE pointing
to a "mutate" file that gets processed. The spawned processes all have
HCP_CONFIG_FILE env-vars that point to the processed output! This creates
confusion, especially with 'docker-compose exec', which bypasses the
container entrypoint. We now use HCP_CONFIG_MUTATE as the input to the
launcher, so that HCP_CONFIG_FILE is only ever set to post-processed host
config.

Signed-off-by: Geoffrey Thorpe <geoff@geoffthorpe.net>